# Fix  `contiguous_split` performance